wip: breaking up table-specific content to a separate module and files for easier maintainability.
