+++ services/ui_backend_service/data/db/tables/__init__.py
@@ -1,0 +1,6 @@
+from .flow import AsyncFlowTablePostgres
+from .run import AsyncRunTablePostgres
+from .step import AsyncStepTablePostgres
+from .task import AsyncTaskTablePostgres
+from .metadata import AsyncMetadataTablePostgres
+from .artifact import AsyncArtifactTablePostgres
