+++ services/ui_backend_service/data/db/tables/artifact.py
@@ -1,0 +1,133 @@
+from .base import AsyncPostgresTable
+from .task import AsyncTaskTablePostgres
+from ..models import ArtifactRow
+from services.data.db_utils import translate_run_key, translate_task_key
+import json
+
+
+class AsyncArtifactTablePostgres(AsyncPostgresTable):
+    artifact_dict = {}
+    run_to_artifact_dict = {}
+    step_to_artifact_dict = {}
+    task_to_artifact_dict = {}
+    current_count = 0
+    _row_type = ArtifactRow
+    table_name = "artifact_v3"
+    task_table_name = AsyncTaskTablePostgres.table_name
+    ordering = ["attempt_id DESC"]
+    keys = ["flow_id", "run_number", "run_id", "step_name", "task_id", "task_name", "name", "location",
+            "ds_type", "sha", "type", "content_type", "user_name", "attempt_id", "ts_epoch", "tags", "system_tags"]
+    primary_keys = ["flow_id", "run_number",
+                    "step_name", "task_id", "attempt_id", "name"]
+    select_columns = keys
+    _command = """
+    CREATE TABLE {0} (
+        flow_id VARCHAR(255) NOT NULL,
+        run_number BIGINT NOT NULL,
+        run_id VARCHAR(255),
+        step_name VARCHAR(255) NOT NULL,
+        task_id BIGINT NOT NULL,
+        task_name VARCHAR(255),
+        name VARCHAR(255) NOT NULL,
+        location VARCHAR(255) NOT NULL,
+        ds_type VARCHAR(255) NOT NULL,
+        sha VARCHAR(255),
+        type VARCHAR(255),
+        content_type VARCHAR(255),
+        user_name VARCHAR(255),
+        attempt_id SMALLINT NOT NULL,
+        ts_epoch BIGINT NOT NULL,
+        tags JSONB,
+        system_tags JSONB,
+        PRIMARY KEY(flow_id, run_number, step_name, task_id, attempt_id, name)
+    )
+    """.format(
+        table_name
+    )
+
+    async def add_artifact(
+        self,
+        flow_id,
+        run_number,
+        run_id,
+        step_name,
+        task_id,
+        task_name,
+        name,
+        location,
+        ds_type,
+        sha,
+        type,
+        content_type,
+        user_name,
+        attempt_id,
+        tags,
+        system_tags,
+    ):
+        dict = {
+            "flow_id": flow_id,
+            "run_number": str(run_number),
+            "run_id": run_id,
+            "step_name": step_name,
+            "task_id": str(task_id),
+            "task_name": task_name,
+            "name": name,
+            "location": location,
+            "ds_type": ds_type,
+            "sha": sha,
+            "type": type,
+            "content_type": content_type,
+            "user_name": user_name,
+            "attempt_id": str(attempt_id),
+            "tags": json.dumps(tags),
+            "system_tags": json.dumps(system_tags),
+        }
+        return await self.create_record(dict)
+
+    async def get_artifacts_in_runs(self, flow_id: str, run_id: int):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+        }
+        return await self.get_records(filter_dict=filter_dict,
+                                      ordering=self.ordering)
+
+    async def get_artifact_in_steps(self, flow_id: str, run_id: int, step_name: str):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+            "step_name": step_name,
+        }
+        return await self.get_records(filter_dict=filter_dict,
+                                      ordering=self.ordering)
+
+    async def get_artifact_in_task(
+        self, flow_id: str, run_id: int, step_name: str, task_id: int
+    ):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        task_id_key, task_id_value = translate_task_key(task_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+            "step_name": step_name,
+            task_id_key: task_id_value,
+        }
+        return await self.get_records(filter_dict=filter_dict,
+                                      ordering=self.ordering)
+
+    async def get_artifact(
+        self, flow_id: str, run_id: int, step_name: str, task_id: int, name: str
+    ):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        task_id_key, task_id_value = translate_task_key(task_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+            "step_name": step_name,
+            task_id_key: task_id_value,
+            '"name"': name,
+        }
+        return await self.get_records(filter_dict=filter_dict,
+                                      fetch_single=True, ordering=self.ordering)
