+++ services/ui_backend_service/data/db/tables/flow.py
@@ -1,0 +1,40 @@
+from .base import AsyncPostgresTable
+from ..models import FlowRow
+import json
+
+
+class AsyncFlowTablePostgres(AsyncPostgresTable):
+    flow_dict = {}
+    table_name = "flows_v3"
+    keys = ["flow_id", "user_name", "ts_epoch", "tags", "system_tags"]
+    primary_keys = ["flow_id"]
+    select_columns = keys
+    join_columns = []
+    _command = """
+    CREATE TABLE {0} (
+        flow_id VARCHAR(255) PRIMARY KEY,
+        user_name VARCHAR(255),
+        ts_epoch BIGINT NOT NULL,
+        tags JSONB,
+        system_tags JSONB
+    )
+    """.format(
+        table_name
+    )
+    _row_type = FlowRow
+
+    async def add_flow(self, flow: FlowRow):
+        dict = {
+            "flow_id": flow.flow_id,
+            "user_name": flow.user_name,
+            "tags": json.dumps(flow.tags),
+            "system_tags": json.dumps(flow.system_tags),
+        }
+        return await self.create_record(dict)
+
+    async def get_flow(self, flow_id: str):
+        filter_dict = {"flow_id": flow_id}
+        return await self.get_records(filter_dict=filter_dict, fetch_single=True)
+
+    async def get_all_flows(self):
+        return await self.get_records()
