+++ services/ui_backend_service/data/db/tables/metadata.py
@@ -1,0 +1,90 @@
+from .base import AsyncPostgresTable
+from .task import AsyncTaskTablePostgres
+from ..models import MetadataRow
+from services.data.db_utils import translate_run_key, translate_task_key
+import json
+
+
+class AsyncMetadataTablePostgres(AsyncPostgresTable):
+    metadata_dict = {}
+    run_to_metadata_dict = {}
+    _current_count = 0
+    _row_type = MetadataRow
+    table_name = "metadata_v3"
+    task_table_name = AsyncTaskTablePostgres.table_name
+    keys = ["flow_id", "run_number", "run_id", "step_name", "task_id", "task_name", "id",
+            "field_name", "value", "type", "user_name", "ts_epoch", "tags", "system_tags"]
+    primary_keys = ["flow_id", "run_number",
+                    "step_name", "task_id", "field_name"]
+    select_columns = keys
+    _command = """
+    CREATE TABLE {0} (
+        flow_id VARCHAR(255),
+        run_number BIGINT NOT NULL,
+        run_id VARCHAR(255),
+        step_name VARCHAR(255) NOT NULL,
+        task_name VARCHAR(255),
+        task_id BIGINT NOT NULL,
+        id BIGSERIAL NOT NULL,
+        field_name VARCHAR(255) NOT NULL,
+        value TEXT NOT NULL,
+        type VARCHAR(255) NOT NULL,
+        user_name VARCHAR(255),
+        ts_epoch BIGINT NOT NULL,
+        tags JSONB,
+        system_tags JSONB,
+        PRIMARY KEY(id, flow_id, run_number, step_name, task_id, field_name)
+    )
+    """.format(
+        table_name
+    )
+
+    async def add_metadata(
+        self,
+        flow_id,
+        run_number,
+        run_id,
+        step_name,
+        task_id,
+        task_name,
+        field_name,
+        value,
+        type,
+        user_name,
+        tags,
+        system_tags,
+    ):
+        dict = {
+            "flow_id": flow_id,
+            "run_number": str(run_number),
+            "run_id": run_id,
+            "step_name": step_name,
+            "task_id": str(task_id),
+            "task_name": task_name,
+            "field_name": field_name,
+            "value": value,
+            "type": type,
+            "user_name": user_name,
+            "tags": json.dumps(tags),
+            "system_tags": json.dumps(system_tags),
+        }
+        return await self.create_record(dict)
+
+    async def get_metadata_in_runs(self, flow_id: str, run_id: str):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        filter_dict = {"flow_id": flow_id,
+                       run_id_key: run_id_value}
+        return await self.get_records(filter_dict=filter_dict)
+
+    async def get_metadata(
+        self, flow_id: str, run_id: int, step_name: str, task_id: str
+    ):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        task_id_key, task_id_value = translate_task_key(task_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+            "step_name": step_name,
+            task_id_key: task_id_value,
+        }
+        return await self.get_records(filter_dict=filter_dict)
