+++ services/ui_backend_service/data/db/tables/step.py
@@ -1,0 +1,61 @@
+from .base import AsyncPostgresTable
+from ..models import StepRow
+from .run import AsyncRunTablePostgres
+from services.data.db_utils import translate_run_key
+import json
+
+
+class AsyncStepTablePostgres(AsyncPostgresTable):
+    step_dict = {}
+    run_to_step_dict = {}
+    _row_type = StepRow
+    table_name = "steps_v3"
+    keys = ["flow_id", "run_number", "run_id", "step_name",
+            "user_name", "ts_epoch", "tags", "system_tags"]
+    primary_keys = ["flow_id", "run_number", "step_name"]
+    select_columns = keys
+    run_table_name = AsyncRunTablePostgres.table_name
+    _command = """
+    CREATE TABLE {0} (
+        flow_id VARCHAR(255) NOT NULL,
+        run_number BIGINT NOT NULL,
+        run_id VARCHAR(255),
+        step_name VARCHAR(255) NOT NULL,
+        user_name VARCHAR(255),
+        ts_epoch BIGINT NOT NULL,
+        tags JSONB,
+        system_tags JSONB,
+        PRIMARY KEY(flow_id, run_number, step_name),
+        FOREIGN KEY(flow_id, run_number) REFERENCES {1} (flow_id, run_number),
+        UNIQUE(flow_id, run_id, step_name)
+    )
+    """.format(
+        table_name, run_table_name
+    )
+
+    async def add_step(self, step_object: StepRow):
+        dict = {
+            "flow_id": step_object.flow_id,
+            "run_number": str(step_object.run_number),
+            "run_id": step_object.run_id,
+            "step_name": step_object.step_name,
+            "user_name": step_object.user_name,
+            "tags": json.dumps(step_object.tags),
+            "system_tags": json.dumps(step_object.system_tags),
+        }
+        return await self.create_record(dict)
+
+    async def get_steps(self, flow_id: str, run_id: str):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        filter_dict = {"flow_id": flow_id,
+                       run_id_key: run_id_value}
+        return await self.get_records(filter_dict=filter_dict)
+
+    async def get_step(self, flow_id: str, run_id: str, step_name: str):
+        run_id_key, run_id_value = translate_run_key(run_id)
+        filter_dict = {
+            "flow_id": flow_id,
+            run_id_key: run_id_value,
+            "step_name": step_name,
+        }
+        return await self.get_records(filter_dict=filter_dict, fetch_single=True)
